+++ Interfaces/Ball.py
@@ -2,36 +2,6 @@ import numpy as np
 import threading, multiprocessing, struct, time
 
 from core.Interface import *
-from Interfaces.RPProbe import *
-
-class VRProbe(RPProbe):
-    channels = {'odor': {1: 19, 2: 16, 3: 26, 4: 20},
-                'liquid': {1: 22},
-                'lick': {1: 17},
-                'sync': {'in': 21},
-                'running': 20}
-    pwm = dict()
-
-    def start_odor(self, channels, dutycycle=50, frequency=20):
-        self.frequency = frequency
-        self.odor_channels = channels
-        for channel in channels:
-            self.pwm[channel] = self.GPIO.PWM(self.channels['odor'][channel], self.frequency)
-            self.pwm[channel].ChangeFrequency(self.frequency)
-            self.pwm[channel].start(dutycycle)
-
-    def update_odor(self, dutycycles):  # for 2D olfactory setup
-        for channel, dutycycle in zip(self.odor_channels, dutycycles):
-            self.pwm[channel].ChangeDutyCycle(dutycycle)
-
-    def stop_odor(self):
-        for channel in self.odor_channels:
-            self.pwm[channel].stop()
-
-    def cleanup(self):
-        for channel in self.odor_channels:
-            self.pwm[channel].stop()
-        super().cleanup()
 
 class Ball(Interface):
     def __init__(self, exp, ball_radius=0.125):
