+++ Interfaces/PCPorts.py
@@ -1,9 +1,9 @@
 from core.Interface import *
 
-class PCProbe(Interface):
+class PCPorts(Interface):
 
     def __init__(self, **kwargs):
-        super(PCProbe, self).__init__(**kwargs)
+        super(PCPorts, self).__init__(**kwargs)
         import serial
         serial_port = 'COM1'
         self.serial = serial.serial_for_url(serial_port)
